# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the project",
        "version": "1.0"
    },
    "paths": {
...
        "/api/pilotos": {
            "get": {
                "summary": "Get list of pilotos",
                "operationId": "499d47088082c9c7c898e6efd0e78799",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new piloto",
                "operationId": "10a298ab55b9dba9c2938701e43a7ca5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "FechaNacimiento",
                                    "Pais",
                                    "EquipoId"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Piloto A"
                                    },
                                    "FechaNacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "Pais": {
                                        "type": "string",
                                        "example": "Pais A"
                                    },
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/pilotos/{id}": {
            "get": {
                "summary": "Get piloto by ID",
                "operationId": "8fc3ee594afb9695e7c0c2e2156d9ed6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update piloto by ID",
                "operationId": "aa4e43062a291c860e18c348570586b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Nombre",
                                    "FechaNacimiento",
                                    "Pais",
                                    "EquipoId"
                                ],
                                "properties": {
                                    "Nombre": {
                                        "type": "string",
                                        "example": "Piloto A"
                                    },
                                    "FechaNacimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-01-01"
                                    },
                                    "Pais": {
                                        "type": "string",
                                        "example": "Pais A"
                                    },
                                    "EquipoId": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete piloto by ID",
                "operationId": "107a1a81ae0e8692c06a1443f00dd0ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/puntos": {
            "get": {
                "summary": "Get list of puntos",
                "operationId": "7478a3dd0eeee60258baa83dfa21915c",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "summary": "Create a new punto",
                "operationId": "92295c3ea74eb9490fbff2c117017d16",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Posicion",
                                    "Puntos"
                                ],
                                "properties": {
                                    "Posicion": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Puntos": {
                                        "type": "integer",
                                        "example": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/puntos/{id}": {
            "get": {
                "summary": "Get punto by ID",
                "operationId": "d9eb99d147cbbde81781319782c82e5b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "summary": "Update punto by ID",
                "operationId": "f722acb63257774327b17d14c42dc48d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Posicion",
                                    "Puntos"
                                ],
                                "properties": {
                                    "Posicion": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "Puntos": {
                                        "type": "integer",
                                        "example": 25
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "summary": "Delete punto by ID",
                "operationId": "529d7cfae17f06accd74413df5b5386d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}